+++ src/lib.rs
@@ -1,4 +1,4 @@
-//! Allows to read the current temperature from the TSIC 306
+//! Allows to read the current temperature from the TSIC temperature sensors.
 //!
 //! Note that most of this code is ported and heavily modified from C
 //! to rust using the code found in [arduino-tsic](https://github.com/Schm1tz1/arduino-tsic)
@@ -10,10 +10,30 @@
 //!
 //! ## Usage
 //!
+//! Please see the comments on both the `with_vdd_control` and `without_vdd_control` constructors for
+//! their usage and upsides/downsides.
+//!
+//! If the library should control both the signal and the vdd pins (recommended):
+//!
 //! ```ignore
 //! use tsic::{SensorType, Tsic};
 //!
-//! let sensor = Tsic::new(SensorType::Tsic306, /* your hal pin */);
+//! let sensor = Tsic::with_vdd_control(SensorType::Tsic306, /* your hal signal input pin */, /* your vdd output pin */);
+//!
+//! let mut delay = /* your hal delay */();
+//!
+//! match sensor.read(&mut delay) {
+//!   Ok(t) => defmt::info!("Temp is: {:f32}", t.as_celsius()),
+//!   Err(e) => defmt::warn!("Getting sensor data failed: {:?}", e),
+//! };
+//! ```
+//!
+//! If the library should just control the signal pin:
+//!
+//! ```ignore
+//! use tsic::{SensorType, Tsic};
+//!
+//! let sensor = Tsic::without_vdd_control(SensorType::Tsic306, /* your hal signal input pin */);
 //!
 //! let mut delay = /* your hal delay */();
 //!
@@ -29,27 +49,69 @@
 
 use core::time::Duration;
 use embedded_hal::blocking::delay::DelayUs;
-use embedded_hal::digital::v2::InputPin;
+use embedded_hal::digital::v2::{InputPin, OutputPin};
 
 /// The spec defines the sample rate as 128kHz, which is 7.8 microseconds. Since
 /// we can only sleep for a round number of micros, 8 micros should be close enough.
 static STROBE_SAMPLING_RATE: Duration = Duration::from_micros(8);
 
+/// The spec says that after power up, the sensor will transmit the first reading after
+/// 65-85ms. Let's play it safe and wait 50 millis after powering up.
+static VDD_POWER_UP_DELAY: Duration = Duration::from_millis(50);
+
 /// The `Tsic` struct is the main entry point when trying to get a temperature reading from a
 /// TSIC 306 sensor.
-pub struct Tsic<I: InputPin> {
+pub struct Tsic<I: InputPin, O: OutputPin> {
     /// Right now the sensor type is unused since we only support one, but it provides a forward
     /// compatible API in case we add support for more in the future.
     _sensor_type: SensorType,
-    pin: I,
+    signal_pin: I,
+    vdd_pin: Option<O>,
 }
 
-impl<I: InputPin> Tsic<I> {
-    /// Creates a new `Tsic` sensor wrapper and binds it to the input pin given.
-    pub fn new(sensor_type: SensorType, pin: I) -> Self {
+impl<I: InputPin> Tsic<I, DummyOutputPin> {
+    /// Constructs a new `Tsic` without explicit control over the voltage (VDD) pin.
+    ///
+    /// Use this construction method if you either want to manage the power of your
+    /// sensor externally or have it on a permanent voltage connection. Usually in
+    /// this case only the signal pin of the sensor is attached to a GPIO pin of
+    /// your board.
+    ///
+    /// *IMPORTANT*: While this sounds like the simpler method, I recommend using
+    /// the `with_vdd_control` constructor and also attach the VDD pin of the sensor
+    /// to your board. This will reduce the risk of accidentially performing a reading
+    /// during the actual temperature transmission. If you still want to use it this
+    /// way, you probably want to consider retrying on transient failures when executing
+    /// the `read` operation.
+    pub fn without_vdd_control(sensor_type: SensorType, signal_pin: I) -> Self {
         Self {
             _sensor_type: sensor_type,
-            pin,
+            signal_pin,
+            vdd_pin: None,
+        }
+    }
+}
+
+impl<I: InputPin, O: OutputPin> Tsic<I, O> {
+    /// Constructs a new `Tsic` with explicit control over the voltage (VDD) pin.
+    ///
+    /// Use this method if you want the library to control the voltage (VDD) pin of the
+    /// sensor as well.
+    ///
+    /// This is the recommended approach because it saves power and it makes
+    /// sure that the readings are very consistent (we do not run the risk of trying to
+    /// perform a reading while one is already in-progress, leading to error values).
+    ///
+    /// Usually you need to assign another GPIO pin as an output pin which can drive around
+    /// 3V in high state (see the datasheet for more info), and then the library will control
+    /// the power up, initial delay, reading and power down for you transparently. Of course,
+    /// you can also use the `without_vdd_control` constructor if you want more manual control
+    /// or if you have the sensor on permanent power.
+    pub fn with_vdd_control(sensor_type: SensorType, signal_pin: I, vdd_pin: O) -> Self {
+        Self {
+            _sensor_type: sensor_type,
+            signal_pin,
+            vdd_pin: Some(vdd_pin),
         }
     }
 
@@ -58,11 +120,55 @@ impl<I: InputPin> Tsic<I> {
     /// Note that the passed in `Delay` from the HAL needs to be aquired outside of
     /// this struct and passed in as mutable, because to aquire correct data from the
     /// sensor the code needs to pause for a certain amount of microseconds.
-    pub fn read<D: DelayUs<u8>>(&self, delay: &mut D) -> Result<Temperature, TsicError> {
-        let first_packet = self.read_packet(delay)?;
-        let second_packet = self.read_packet(delay)?;
+    ///
+    /// In case there is an error during the read phase and if the `Tsic` has been constructed
+    /// to manage the VDD pin as well, it will try to shut it down in a best-effort manner as
+    /// well.
+    pub fn read<D: DelayUs<u16>>(&mut self, delay: &mut D) -> Result<Temperature, TsicError> {
+        self.maybe_power_up_sensor(delay)?;
+
+        let first_packet = match self.read_packet(delay) {
+            Ok(packet) => packet,
+            Err(err) => {
+                self.maybe_power_down_sensor().ok();
+                return Err(err);
+            }
+        };
+
+        let second_packet = match self.read_packet(delay) {
+            Ok(packet) => packet,
+            Err(err) => {
+                self.maybe_power_down_sensor().ok();
+                return Err(err);
+            }
+        };
+
+        self.maybe_power_down_sensor()?;
 
         Ok(Temperature::new(first_packet, second_packet))
+    }
+
+    /// Handle VDD pin power up if set during construction.
+    ///
+    /// If we are managing the VDD pin for the user, we need to power up the sensor and then
+    /// apply an initial delay before the reading can continue.
+    fn maybe_power_up_sensor<D: DelayUs<u16>>(&mut self, delay: &mut D) -> Result<(), TsicError> {
+        if let Some(ref mut pin) = self.vdd_pin {
+            pin.set_high().map_err(|_| TsicError::PinWriteError)?;
+            delay.delay_us(VDD_POWER_UP_DELAY.as_micros() as u16);
+        }
+        Ok(())
+    }
+
+    /// Handle VDD pin power down if set during construction.
+    ///
+    /// If we are managing the VDD pin for the user, at the end of the measurement we need
+    /// to power it down at the end as well.
+    fn maybe_power_down_sensor(&mut self) -> Result<(), TsicError> {
+        if let Some(ref mut pin) = self.vdd_pin {
+            pin.set_low().map_err(|_| TsicError::PinWriteError)?;
+        }
+        Ok(())
     }
 
     /// Reads the bits off of the sensor port based on the ZACWire protocol.
@@ -84,10 +190,10 @@ impl<I: InputPin> Tsic<I> {
     ///
     /// See https://www.ist-ag.com/sites/default/files/ATTSic_E.pdf for
     /// the full document.
-    fn read_packet<D: DelayUs<u8>>(&self, delay: &mut D) -> Result<Packet, TsicError> {
+    fn read_packet<D: DelayUs<u16>>(&self, delay: &mut D) -> Result<Packet, TsicError> {
         self.wait_until_low()?;
 
-        let strobe_len = self.strobe_len(delay)?.as_micros() as u8;
+        let strobe_len = self.strobe_len(delay)?.as_micros() as u16;
 
         let mut packet_bits: u16 = 0;
 
@@ -114,13 +220,13 @@ impl<I: InputPin> Tsic<I> {
     /// read attempt.
     ///
     /// The strobe length should be around 60 microseconds.
-    fn strobe_len<D: DelayUs<u8>>(&self, delay: &mut D) -> Result<Duration, TsicError> {
+    fn strobe_len<D: DelayUs<u16>>(&self, delay: &mut D) -> Result<Duration, TsicError> {
         let sampling_rate = STROBE_SAMPLING_RATE.as_micros();
 
         let mut strobe_len = 0;
         while self.is_low()? {
             strobe_len += sampling_rate;
-            delay.delay_us(sampling_rate as u8);
+            delay.delay_us(sampling_rate as u16);
         }
 
         Ok(Duration::from_micros(strobe_len as u64))
@@ -128,12 +234,16 @@ impl<I: InputPin> Tsic<I> {
 
     /// Checks if the pin is currently in a high state.
     fn is_high(&self) -> Result<bool, TsicError> {
-        self.pin.is_high().map_err(|_| TsicError::PinReadError)
+        self.signal_pin
+            .is_high()
+            .map_err(|_| TsicError::PinReadError)
     }
 
     /// Checks if the pin is currently in a low state.
     fn is_low(&self) -> Result<bool, TsicError> {
-        self.pin.is_low().map_err(|_| TsicError::PinReadError)
+        self.signal_pin
+            .is_low()
+            .map_err(|_| TsicError::PinReadError)
     }
 
     /// Returns only once the pin is in a low state.
@@ -158,8 +268,11 @@ pub enum TsicError {
     /// read might resolve the error.
     ParityCheckFailed,
 
-    /// Failed to read the high/low state of the pin.
+    /// Failed to read the high/low state of signal the pin.
     PinReadError,
+
+    /// Failed to set the high/low state of the vdd pin.
+    PinWriteError,
 }
 
 /// Represents a single temperature reading from the TSIC 306 sensor.
@@ -218,4 +331,22 @@ impl Packet {
 pub enum SensorType {
     /// Use this variant if you use the TSic 306 sensor.
     Tsic306,
+}
+
+/// This `OutputPin` is used to satisfy the generics when no explicit pin is provided.
+///
+/// Note that you do not want to use this struct, I just couldn' figure out a much
+/// better way right now, but hopefully it will go away at some point.
+pub struct DummyOutputPin {}
+
+impl OutputPin for DummyOutputPin {
+    type Error = ();
+
+    fn set_low(&mut self) -> Result<(), Self::Error> {
+        Ok(())
+    }
+
+    fn set_high(&mut self) -> Result<(), Self::Error> {
+        Ok(())
+    }
 }