+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+ - Added support for VDD management. This is now the recommended approach.
  - Added `SensorType` so that we leave the door open to more sensor support in the future.
 
 ### Changed
 
+ - Changed `new` to both `without_vdd_control` or `with_vdd_control` with the new VDD management.
  - Renamed `TsicReadError` to `TsicError` to make it more generic.
 
 ## v0.1.0 - 2018-09-28